apps/dialogue-generator/requirements.txt
fastapi
uvicorn
dapr
dapr-ext-grpc
langchain-community